requirements.txt
PySwitchbot==0.14.0